requirements.txt
streamlit>=1.45.0
openai>=1.82.0
anthropic>=0.52.0
google-genai>=1.14.0
openpyxl>=3.1.0
Pillow>=11.0.0
pdf2image>=1.17.0
python-dotenv>=1.0.0
pytest>=8.3.0
pytest-cov>=6.0.0
